# server/swagger.yaml
openapi: 3.0.0
info:
  title: sms
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /classses/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - classses
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - classses
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/classses"
    post:
      summary: CRUD ACTION update
      tags:
        - classses
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/classses"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/classses"
  /classses/findByschool_id/{key}:
    get:
      summary: CRUD ACTION findByschool_id
      tags:
        - classses
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/classses"
    
  #-----  gender  -----
...
  /gender/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - gender
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - gender
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/gender"
    post:
      summary: CRUD ACTION update
      tags:
        - gender
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/gender"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/gender"
    
  #-----  religion  -----
...
  /roles/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - roles
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - roles
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/roles"
    post:
      summary: CRUD ACTION update
      tags:
        - roles
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/roles"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/roles"
    
  #-----  school  -----
...
  /section/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - section
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - section
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/section"
    post:
      summary: CRUD ACTION update
      tags:
        - section
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/section"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/section"
  /section/findByacademic_id/{key}:
    get:
      summary: CRUD ACTION findByacademic_id
      tags:
        - section
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/section"
  /section/findByclasss_id/{key}:
    get:
      summary: CRUD ACTION findByclasss_id
      tags:
        - section
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/section"
  /section/findByschool_id/{key}:
    get:
      summary: CRUD ACTION findByschool_id
      tags:
        - section
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/section"
    
  #-----  student  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    classses:
      properties:
        _id:
          type: string
        active_status:
          type: boolean
        name:
          type: string
      required:
        -  _id
    gender:
      properties:
        _id:
          type: string
        name:
          type: string
        short_name:
          type: string
      required:
        -  _id
...
    roles:
      properties:
        _id:
          type: string
        active_status:
          type: boolean
        name:
          type: string
        type:
          type: string
      required:
        -  _id
...
    section:
      properties:
        _id:
          type: string
        active_status:
          type: boolean
        name:
          type: string
      required:
        -  _id
...